lc_projects/requirements.txt
langchain-openai
langchain
python-dotenv
langchain_community
langchain_core
streamlit
altair
langchain_ollama
langchain_groq
pypdf
faiss-cpu